fix(a11y): set aria-hidden on prev/next link text
Since these have a label we need to hide the text content which is used
as the label. Right now these announce both the `aria-label` and the
text content.

It could be tempting to wrap the arrows with a span with
`aria-hidden=true` on it and then add a visually hidden span with "Page"
in it, but that would probably suffer from screen readers fragmenting
the text.